+++ edji-frontend/src/hooks.server.js
@@ -1,4 +1,13 @@
-export function handle({ event, resolve }) {
-	event.locals.token = event.cookies.get('token');
-	return resolve(event);
+import { handleSession } from 'svelte-kit-cookie-session';
+import {SECRET_KEY} from '$env/static/private'
+
+function redirect(location, body) {
+    return new Response(body, {
+        status: 303,
+        headers: { location }
+    });
 }
+
+export const handle = handleSession({
+	secret: SECRET_KEY,
+});