+++ edji-frontend/src/routes/+page.svelte
@@ -15,9 +15,6 @@
 	<section>
 		<h2>A no-fuss, full-featured Restful back-end.</h2>
 	</section>
-	<div class="bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded relative" role="alert">
-		This is a standard alert. You can customize its content and even the icon.
-	</div>
 </div>
 <style>