+++ edji-frontend/src/routes/login/+page.server.js
@@ -2,7 +2,7 @@ import { redirect }  from "@sveltejs/kit"
 import { BACKEND } from "$lib/api/edji.js";
 
 export const actions = {
-    default: async ({ request, cookies }) => {
+    default: async ({ request, locals }) => {
         const data = await request.formData();
         const email = data.get("email");
         const password = data.get("password");
@@ -17,11 +17,9 @@ export const actions = {
             })
         });
         const login_result = await result.json();
-        console.log(login_result.token);
         if (login_result.token) {
-            cookies.set("token", login_result.token);
-            console.log("redirecting")
-            throw redirect(307, "/dashboard");
+            await locals.session.set({ token: login_result.token });
+            throw redirect(303, "/dashboard");
         } else {
             return {
                 login_error: "Invalid email or password."