+++ CHANGELOG.md
@@ -5,4 +5,4 @@
 - DOCKER: Fix issue with frontend ports
 - DOCKER: Ignore node_modules (.dockerignore)
 - BUG: Link to API wasn't working. Fixed.
-- Login working
+- Login working - using encrypted token.